合并

+++ clem/type/point.cpp
@@ -10,7 +10,7 @@ Point::Point()
 {
 }
 
-Point::Point(ushort x, ushort y)
+Point::Point(float x, float y)
 		: x(x), y(y)
 {
 }
